Added connection for HanaDB and MySQL
