+++ src/connection_tool/LogSaver.java
@@ -1,3 +1,5 @@
+package connection_tool;
+
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
