+++ src/connection_tool/Provider.java
@@ -1,3 +1,5 @@
+package connection_tool;
+
 public enum Provider {
     ORACLE,
     MYSQL,
