+++ src/connection_tool/connections/DB2Connection.java
@@ -1,45 +1,53 @@
+package connection_tool.connections;
+
+import connection_tool.LogSaver;
+
 import java.util.Properties;
 import java.util.logging.Level;
 
+public class DB2Connection implements IConnection {
 
-public class MySQLConnection implements IConnection {
-
-    private static final String MYSQL_PREFIX = "jdbc:mysql";
+    private static final String DB2_PREFIX = "jdbc:db2://";
     private final String host;
     private final String port;
     private final String databaseName;
     private final String username;
     private final String password;
     private final boolean sslEnabled;
-    private final int timeout;
+    private int timeout;
 
-    public MySQLConnection(Properties properties) {
+
+    public DB2Connection(Properties properties) {
         this.host = properties.getProperty("host");
         this.port = properties.getProperty("port");
         this.databaseName = properties.getProperty("db_name");
         this.sslEnabled = Boolean.parseBoolean(properties.getProperty("ssl"));
         this.username = properties.getProperty("username");
         this.password = properties.getProperty("password");
-        this.timeout = Integer.parseInt(properties.getProperty("timeout"));
-
-
+        this.timeout = 0;
+        try {
+            this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        }catch (NumberFormatException e){
+            System.out.println("Add timeout time to configuration");
+            System.exit(0);
+        }
         LogSaver.appendLog(Level.INFO, "JDBC String: " + getConnectionString()+  "\n" +
                 "User: " + username + "\n" +
                 "Hostname: " + host);
     }
 
     @Override
-    public String getConnectionString() {
-        return MYSQL_PREFIX + host + ":" + port + "/" + databaseName;
+    public String getConnectionString(){
+        return DB2_PREFIX+host + ":" + port +"/" + databaseName;
     }
-
     @Override
-    public Properties getProperties() {
+    public Properties getProperties(){
         var properties = new Properties();
-        properties.put("user", username);
-        properties.put("password", password);
+        properties.setProperty("user", username);
+        properties.setProperty("password", password);
+        properties.setProperty("connection timeout", String.valueOf(timeout));
         if (sslEnabled) {
-            properties.put("useSSL", "true");
+            properties.setProperty("sslConnection", "true");
         }
         return properties;
     }
@@ -53,4 +61,5 @@ public class MySQLConnection implements IConnection {
     public int getTimeout() {
         return timeout;
     }
+
 }
