+++ src/connection_tool/connections/IConnection.java
@@ -1,3 +1,5 @@
+package connection_tool.connections;
+
 import java.util.Properties;
 
 public interface IConnection {
