+++ src/connection_tool/connections/MSQLConnection.java
@@ -1,3 +1,7 @@
+package connection_tool.connections;
+
+import connection_tool.LogSaver;
+
 import java.util.Properties;
 import java.util.logging.Level;
 
@@ -13,11 +17,7 @@ public class MSQLConnection implements IConnection {
 
     private final String username;
     private final String password;
-    private final String authenticationScheme;
-    private final String domain;
-    private final String realm;
-    private final String kdc;
-    private final int timeout;
+    private int timeout;
 
     public MSQLConnection(Properties properties) {
        this.host = properties.getProperty("host");
@@ -26,17 +26,13 @@ public class MSQLConnection implements IConnection {
        this.databaseName = properties.getProperty("db_name");
        this.sslEnabled = Boolean.parseBoolean(properties.getProperty("ssl"));
        this.validateCertificates = Boolean.parseBoolean(properties.getProperty("validate_certificates"));
-       this.authenticationScheme = properties.getProperty("authentication_scheme");
        this.username = properties.getProperty("username");
        this.password = properties.getProperty("password");
-       this.domain  = properties.getProperty("domain");
-       this.realm = properties.getProperty("realm");
-       this.kdc = properties.getProperty("kdc");
-       this.timeout = Integer.parseInt(properties.getProperty("timeout"));
         try {
-            checkProps();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+            this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        }catch (NumberFormatException e){
+            System.out.println("Add timeout time to configuration");
+            System.exit(0);
         }
 
         LogSaver.appendLog(Level.INFO, "JDBC String: " + getConnectionString()+  "\n" +
@@ -47,22 +43,9 @@ public class MSQLConnection implements IConnection {
 
     @Override
     public String getConnectionString(){
-        String connectionString =  SQLSERVER_PREFIX+host + ":" + port;
-        System.out.println(connectionString);
-        return connectionString;
+        return SQLSERVER_PREFIX+host + ":" + port;
     }
 
-    private void checkProps() throws Exception {
-
-        if (authenticationScheme.equals("kerberos")) {
-            if (realm.isBlank() || realm.isEmpty()) {
-                throw new Exception("Realm field can not be empty!");
-            }
-            if (kdc.isEmpty() || kdc.isBlank()) {
-                throw new Exception("KDC can not be empty!");
-            }
-        }
-    }
     @Override
     public Properties getProperties(){
         Properties properties = new Properties();
@@ -72,21 +55,8 @@ public class MSQLConnection implements IConnection {
 
         properties.put("database", databaseName);
         properties.put("instanceName", instanceName);
+        properties.put("loginTimeout", String.valueOf(timeout));
 
-        switch (authenticationScheme){
-            case "basic":
-                break;
-            case "kerberos":
-                properties.put("realm", realm);
-                properties.put("kdc", kdc); //TODO
-                break;
-            case "ntlm":
-                properties.put("domain", domain);
-                break;
-            default:
-                System.out.println("Declare authentication scheme in properties: basic, kerberos, ntlm");
-                System.exit(0);
-        }
         if (sslEnabled){
             properties.put("encrypt", "true");
         }
