+++ src/connection_tool/connections/MySQLConnection.java
@@ -1,50 +1,57 @@
+package connection_tool.connections;
+
+import connection_tool.LogSaver;
+
 import java.util.Properties;
-import java.util.Scanner;
 import java.util.logging.Level;
 
-public class PostgreSQLConnection implements IConnection{
 
-    private static final Scanner scanner = new Scanner(System.in);
-    private static final String POSTGRESQL_PREFIX = "jdbc:postgresql://";
+public class MySQLConnection implements IConnection {
+
+    private static final String MYSQL_PREFIX = "jdbc:mariadb://";
     private final String host;
     private final String port;
     private final String databaseName;
-    private  final String username;
+    private final String username;
     private final String password;
-
     private final boolean sslEnabled;
-    private final int timeout;
+    private int timeout;
 
-
-
-
-    public PostgreSQLConnection(Properties properties) {
-
+    public MySQLConnection(Properties properties) {
         this.host = properties.getProperty("host");
         this.port = properties.getProperty("port");
         this.databaseName = properties.getProperty("db_name");
         this.sslEnabled = Boolean.parseBoolean(properties.getProperty("ssl"));
         this.username = properties.getProperty("username");
         this.password = properties.getProperty("password");
-        this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        try {
+            this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        }catch (NumberFormatException e){
+            System.out.println("Add timeout time to configuration");
+            System.exit(0);
+        }
+
 
         LogSaver.appendLog(Level.INFO, "JDBC String: " + getConnectionString()+  "\n" +
                 "User: " + username + "\n" +
                 "Hostname: " + host);
-        scanner.close();
     }
 
     @Override
-    public String getConnectionString(){
-        return POSTGRESQL_PREFIX+host + ":" + port +"/" + databaseName;
+    public String getConnectionString() {
+        return MYSQL_PREFIX + host + ":" + port + "/" + databaseName;
     }
+
     @Override
-    public Properties getProperties(){
+    public Properties getProperties() {
         var properties = new Properties();
         properties.put("user", username);
         properties.put("password", password);
+        properties.put("connectTimeout", timeout * 1000);
         if (sslEnabled) {
-            properties.put("ssl", "true");
+            properties.put("sslMode", "trust");
+        }else {
+            properties.put("sslMode", "disable");
         }
         return properties;
     }
@@ -58,6 +65,4 @@ public class PostgreSQLConnection implements IConnection{
     public int getTimeout() {
         return timeout;
     }
-
-
 }
