+++ src/connection_tool/connections/PostgreSQLConnection.java
@@ -1,43 +1,64 @@
+package connection_tool.connections;
+
+import connection_tool.LogSaver;
+
 import java.util.Properties;
+import java.util.Scanner;
 import java.util.logging.Level;
 
-public class DB2Connection implements IConnection {
+public class PostgreSQLConnection implements IConnection {
 
-    private static final String DB2_PREFIX = "jdbc:db2://";
+    private static final Scanner scanner = new Scanner(System.in);
+    private static final String POSTGRESQL_PREFIX = "jdbc:postgresql://";
     private final String host;
     private final String port;
     private final String databaseName;
-    private final String username;
+    private  final String username;
     private final String password;
+
     private final boolean sslEnabled;
-    private final int timeout;
+    private int timeout;
 
 
-    public DB2Connection(Properties properties) {
+
+
+    public PostgreSQLConnection(Properties properties) {
+
         this.host = properties.getProperty("host");
         this.port = properties.getProperty("port");
         this.databaseName = properties.getProperty("db_name");
         this.sslEnabled = Boolean.parseBoolean(properties.getProperty("ssl"));
         this.username = properties.getProperty("username");
         this.password = properties.getProperty("password");
-        this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        try {
+            this.timeout = Integer.parseInt(properties.getProperty("timeout"));
+        }catch (NumberFormatException e){
+            System.out.println("Add timeout time to configuration");
+            System.exit(0);
+        }
+
         LogSaver.appendLog(Level.INFO, "JDBC String: " + getConnectionString()+  "\n" +
                 "User: " + username + "\n" +
                 "Hostname: " + host);
+        scanner.close();
     }
 
     @Override
     public String getConnectionString(){
-        return DB2_PREFIX+host + ":" + port +"/" + databaseName;
-  //      return "jdbc:db2://db2DateInstance.lab.dynatrace.org:25001/SAMPLE:sslConnection=true;";
+        if(!port.isBlank())
+            return POSTGRESQL_PREFIX+host + ":" + port +"/" + databaseName;
+        else {
+            return POSTGRESQL_PREFIX + host + "/" + databaseName;
+        }
     }
     @Override
     public Properties getProperties(){
         var properties = new Properties();
         properties.put("user", username);
         properties.put("password", password);
+        properties.put("loginTimeout", timeout);
         if (sslEnabled) {
-            properties.put("sslConnection", "true");
+            properties.put("ssl", "true");
         }
         return properties;
     }
@@ -51,5 +72,6 @@ public class DB2Connection implements IConnection {
     public int getTimeout() {
         return timeout;
     }
+
 
 }
