+++ src/connection_tool/resources/db2.properties
@@ -4,5 +4,5 @@ db_name=SAMPLE
 username=db2inst1
 password=labpass
 ssl=false
-timeout=
+timeout=10
