+++ src/connection_tool/resources/hanadb.properties
@@ -1,0 +1,6 @@
+host=172.18.150.24
+port=39015
+username=SYSTEM
+password=Adlex123
+ssl=true
+timeout=10
