+++ src/connection_tool/resources/msql.properties
@@ -4,10 +4,6 @@ instance_name=
 db_name=
 ssl=true
 validate_certificates=true
-authentication_scheme=basic
 username=sa
 password=Labpass123
-domain=
-realm=
-kdc=
-timeout=
+timeout=10
