+++ src/connection_tool/resources/mysql.properties
@@ -1,0 +1,7 @@
+host=172.18.151.50
+port=3306
+db_name=mysql
+username=piotr
+password=piotr
+ssl=true
+timeout=10
