+++ src/connection_tool/resources/oracle.properties
@@ -1,0 +1,8 @@
+host=ORACLE-OL-SSL
+port=1521
+service_name=
+sid=orcl
+username=sys as sysdba
+password=labpass
+ssl=
+timeout=10
