+++ src/connection_tool/resources/postgresql.properties
@@ -4,4 +4,4 @@ db_name=
 username=postgres
 password=postgres
 ssl=
-timeout=
+timeout=10
