+++ src/connection_tool/resources/snowflake.properties
@@ -1,8 +1,8 @@
 host=
 port=
-db_name=
+db_name=SNOWFLAKE
 schema=
 warehouse=
 username=
 password=
-timeout=
+timeout=10
